fix: extend async audio modules with overridable options
