+++ package/expo-package/src/index.js
@@ -11,6 +11,7 @@ import {
   getPhotos,
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
+  overrideAudioRecordingConfiguration,
   pickDocument,
   pickImage,
   saveFile,
@@ -31,6 +32,7 @@ registerNativeHandlers({
   getPhotos,
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
+  overrideAudioRecordingConfiguration,
   pickDocument,
   pickImage,
   saveFile,
