+++ package/native-package/src/index.js
@@ -12,6 +12,7 @@ import {
   getPhotos,
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
+  overrideAudioRecordingConfiguration,
   pickDocument,
   pickImage,
   saveFile,
@@ -32,6 +33,7 @@ registerNativeHandlers({
   getPhotos,
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
+  overrideAudioRecordingConfiguration,
   pickDocument,
   pickImage,
   saveFile,
